move range-generation functions to arrays._ranges
